# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LeadFlow Marketing Automation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  name?: string;
  creditsRemaining: number;
  creditsUsed: number;
  planId?: string | null;
  planName?: string | null;
  planExpiresAt?: string | null;
  createdAt?: string;
}

export interface DashboardStats {
  creditsRemaining: number;
  creditsUsed: number;
  totalCampaigns: number;
  activeCampaigns: number;
  totalLeads: number;
  emailsSent: number;
  whatsappSent: number;
  planName?: string | null;
  planExpiresAt?: string | null;
}

export type TemplateType = (typeof TemplateType)[keyof typeof TemplateType];

export const TemplateType = {
  email: "email",
  whatsapp: "whatsapp",
} as const;

export interface Template {
  id: number;
  name: string;
  type: TemplateType;
  subject?: string | null;
  body: string;
  createdAt?: string;
  updatedAt?: string;
}

export type CreateTemplateBodyType =
  (typeof CreateTemplateBodyType)[keyof typeof CreateTemplateBodyType];

export const CreateTemplateBodyType = {
  email: "email",
  whatsapp: "whatsapp",
} as const;

export interface CreateTemplateBody {
  name: string;
  type: CreateTemplateBodyType;
  subject?: string | null;
  body: string;
}

export type CampaignStatus =
  (typeof CampaignStatus)[keyof typeof CampaignStatus];

export const CampaignStatus = {
  draft: "draft",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface Campaign {
  id: number;
  name: string;
  niche: string;
  status: CampaignStatus;
  emailSubject?: string | null;
  emailBody?: string | null;
  whatsappMessage?: string | null;
  sendEmail: boolean;
  sendWhatsapp: boolean;
  totalLeads: number;
  emailsSent: number;
  whatsappSent: number;
  createdAt?: string;
  updatedAt?: string;
}

export interface CreateCampaignBody {
  name: string;
  niche: string;
  emailSubject?: string | null;
  emailBody?: string | null;
  whatsappMessage?: string | null;
  sendEmail: boolean;
  sendWhatsapp: boolean;
  templateId?: number | null;
}

export interface ScrapeLeadsBody {
  niche: string;
  maxResults?: number;
}

export interface Lead {
  email?: string | null;
  phone?: string | null;
  website?: string | null;
  name?: string | null;
  source?: string | null;
}

export interface ScrapeLeadsResult {
  leads: Lead[];
  total: number;
  creditsUsed: number;
}

export interface GmailIntegration {
  id?: number | null;
  senderEmail?: string | null;
  isConnected: boolean;
  createdAt?: string | null;
}

export interface SaveGmailBody {
  senderEmail: string;
  appPassword: string;
}

export interface WhatsappStatus {
  isConnected: boolean;
  phone?: string | null;
  sessionExists: boolean;
}

export interface WhatsappQR {
  qrCode?: string | null;
  status: string;
}

export interface BillingPlan {
  id: string;
  name: string;
  price: number;
  currency: string;
  durationDays: number;
  credits: number;
  features: string[];
  isPopular: boolean;
}

export interface Subscription {
  id?: number | null;
  planId?: string | null;
  planName?: string | null;
  status?: string | null;
  expiresAt?: string | null;
  isActive: boolean;
}

export type CreditTransactionType =
  (typeof CreditTransactionType)[keyof typeof CreditTransactionType];

export const CreditTransactionType = {
  credit: "credit",
  debit: "debit",
} as const;

export interface CreditTransaction {
  id: number;
  type: CreditTransactionType;
  amount: number;
  description: string;
  createdAt: string;
}

export type BillingWebhookBody = { [key: string]: unknown };