# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LeadFlow Marketing Automation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: getCampaigns
      tags: [campaigns]
      summary: List all campaigns
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create and launch a campaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign details with lead stats
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /leads/scrape:
    post:
      operationId: scrapeLeads
      tags: [leads]
      summary: Scrape leads from DuckDuckGo for a niche
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeLeadsBody"
      responses:
        "200":
          description: Scraped leads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeLeadsResult"
  /integrations/gmail:
    get:
      operationId: getGmailIntegration
      tags: [integrations]
      summary: Get Gmail integration status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Gmail integration status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GmailIntegration"
    post:
      operationId: saveGmailIntegration
      tags: [integrations]
      summary: Save Gmail sender email and app password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveGmailBody"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GmailIntegration"
  /integrations/whatsapp/status:
    get:
      operationId: getWhatsappStatus
      tags: [integrations]
      summary: Get WhatsApp session status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: WhatsApp status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappStatus"
  /integrations/whatsapp/connect:
    post:
      operationId: connectWhatsapp
      tags: [integrations]
      summary: Initiate WhatsApp connection (returns QR code data)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: QR code for scanning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappQR"
  /integrations/whatsapp/disconnect:
    post:
      operationId: disconnectWhatsapp
      tags: [integrations]
      summary: Disconnect WhatsApp session
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Disconnected
...
  /billing/subscription:
    get:
      operationId: getSubscription
      tags: [billing]
      summary: Get current user subscription
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /billing/webhook:
    post:
      operationId: billingWebhook
      tags: [billing]
      summary: Payment webhook for verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        niche:
          type: string
        status:
          type: string
          enum: [draft, running, completed, failed]
        emailSubject:
          type: string
          nullable: true
        emailBody:
          type: string
          nullable: true
        whatsappMessage:
          type: string
          nullable: true
        sendEmail:
          type: boolean
        sendWhatsapp:
          type: boolean
        totalLeads:
          type: integer
        emailsSent:
          type: integer
        whatsappSent:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - niche
        - status
        - sendEmail
        - sendWhatsapp
        - totalLeads
        - emailsSent
        - whatsappSent
    CreateCampaignBody:
      type: object
      properties:
        name:
          type: string
        niche:
          type: string
        emailSubject:
          type: string
          nullable: true
        emailBody:
          type: string
          nullable: true
        whatsappMessage:
          type: string
          nullable: true
        sendEmail:
          type: boolean
        sendWhatsapp:
          type: boolean
        templateId:
          type: integer
          nullable: true
      required:
        - name
        - niche
        - sendEmail
        - sendWhatsapp
    ScrapeLeadsBody:
      type: object
      properties:
        niche:
          type: string
        maxResults:
          type: integer
      required:
        - niche
    ScrapeLeadsResult:
      type: object
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        creditsUsed:
          type: integer
      required:
        - leads
        - total
        - creditsUsed
    Lead:
      type: object
      properties:
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
      required: []
    GmailIntegration:
      type: object
      properties:
        id:
          type: integer
          nullable: true
        senderEmail:
          type: string
          nullable: true
        isConnected:
          type: boolean
        createdAt:
          type: string
          nullable: true
      required:
        - isConnected
    SaveGmailBody:
      type: object
      properties:
        senderEmail:
          type: string
        appPassword:
          type: string
      required:
        - senderEmail
        - appPassword
    WhatsappStatus:
      type: object
      properties:
        isConnected:
          type: boolean
        phone:
          type: string
          nullable: true
        sessionExists:
          type: boolean
      required:
        - isConnected
        - sessionExists
    WhatsappQR:
      type: object
      properties:
        qrCode:
          type: string
          nullable: true
        status:
          type: string
      required:
        - status
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
          nullable: true
        planId:
          type: string
          nullable: true
        planName:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        expiresAt:
          type: string
          nullable: true
        isActive:
          type: boolean
      required:
        - isActive
...