# lib/api-zod/src/generated/types/scrapeLeadsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LeadFlow Marketing Automation API
 * OpenAPI spec version: 0.1.0
 */

export interface ScrapeLeadsBody {
  niche: string;
  maxResults?: number;
}