# lib/api-zod/src/generated/types/scrapeLeadsResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LeadFlow Marketing Automation API
 * OpenAPI spec version: 0.1.0
 */
import type { Lead } from "./lead";

export interface ScrapeLeadsResult {
  leads: Lead[];
  total: number;
  creditsUsed: number;
}