# lib/api-zod/src/generated/types/whatsappStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LeadFlow Marketing Automation API
 * OpenAPI spec version: 0.1.0
 */

export interface WhatsappStatus {
  isConnected: boolean;
  phone?: string | null;
  sessionExists: boolean;
}